Move Blackjack MoveChecker to the bot file.
