+++ Managers/SessionManagers/blackjack_bot.py
@@ -1,4 +1,3 @@
-from Blackjack.move_checker import BlackjackMoveChecker
 from Blackjack.blackjack_executor import BlackjackExecutor
 from Blackjack.join_timer import BlackjackJoinTimer
 from Core.core_game_class import GameCore
@@ -81,3 +80,48 @@ class BlackjackInitializer(GameInitializer):
 
     def _get_game_to_create(self, ctx) -> GameCore:
         return BlackjackExecutor(self.bot, ctx)
+
+
+class BlackjackMoveChecker:
+
+    """
+    Checks if your Blackjack move is legal.
+    """
+
+    def __init__(self, bot, game):
+        self.bot = bot
+        self.game = game
+
+    async def perform_action(self, ctx, action_to_perform: str):
+        user = ctx.message.author
+        can_make_move = await self._can_make_move(user)
+        if can_make_move:
+            action_list = self.get_actions()
+            await action_list[action_to_perform]()
+
+    def get_actions(self):
+        return {
+                "hit": self.game.hit,
+                "stand": self.game.stand_current_hand,
+                "split": self.game.attempt_split,
+                "doubledown": self.game.attempt_double_down
+                }
+
+    async def _can_make_move(self, user) -> bool:
+        move_error = self._check_move_error(user)
+        if move_error:
+            await self.bot.say(move_error)
+        else:
+            return True
+
+    def _check_move_error(self, user) -> any:
+        error = False
+        if not self._is_valid_turn(user):
+            error = "It's not your turn. Please wait."
+        elif not self.game.in_progress:
+            error = "The game is not underway yet."
+        return error
+
+    def _is_valid_turn(self, user) -> bool:
+        first_in_queue = self.game.get_current_player().user
+        return user is first_in_queue
